* master: (184 commits)
  CF system spec creates RDS resources first
  fix microbosh system spec vpc/rds outfile paths
  fix microBOSH spec after deployment generator was changed to take two arguments
  enhance cpi tests for aws plugin and error message
  microBOSH system spec creates BAT deployment directory just when needed
  HTTP plays well with a proxy environment
  add retryable functionality to Bosh::Common module
  Move generate_build_number into CompiledPackage, add specs for it
  Added logging for uploading compiled packages to global cache
  Global cache key respects dependent package versions
  No need to check if it is a block device now
  Add some changes to pull request #59
  fix randomly failing aws boostrap cli spec
  move bosh_agent/misc/temcell/uild2 to the top level directory
  Added some tests and made changes to handle \n in the filename if any while searching for cdrom device
  vendor retryable gem
  Add Retryable gem
  added bosh-stemcell-vsphere 0.8.0 (test)
  update stemcell build for 3.0 kernel
  Use if_p in director template so it doesn't blow up when global_blobstore is missing
  ...

+++ blobstore_client/lib/blobstore_client/local_client.rb
@@ -1,7 +1,5 @@
 # Copyright (c) 2009-2012 VMware, Inc.
 
-require "uuidtools"
-
 module Bosh
   module Blobstore
     class LocalClient < BaseClient
@@ -14,9 +12,10 @@ module Bosh
         FileUtils.mkdir_p(@blobstore_path) unless File.directory?(@blobstore_path)
       end
 
-      def create_file(file)
-        id = UUIDTools::UUID.random_create.to_s
+      def create_file(id, file)
+        id ||= UUIDTools::UUID.random_create.to_s
         dst = File.join(@blobstore_path, id)
+        raise BlobstoreError, "object id #{id} is already in use" if File.exist?(dst)
         File.open(dst, 'w') do |fh|
           until file.eof?
             fh.write(file.read(CHUNK_SIZE))
